Initial steps for variable allocation.
Added separate classes for function head, body, and return scopes.

Added separate Value classes corresponding to these, and Variable
always provide an implicit Value if it got no explicit pivot Value.

FunctionScope now contains all three child scopes for clarity.

FunctionReturnValue gets a pointer to the return scope, so it will know
where to put the return value.

The function definition injects an implicit return variable, based
on the type specified.

Added Declaration::allocate method.

Added Scope::reserve method, handling both up and downward growing scopes.

Fixed some -Wextra complaints.

+++ arch/ia32.cpp
@@ -169,7 +169,7 @@ void IA32::done(std::string filename) {
         case DEF_EXPORT_CODE:
             switch (r.type) {
                 case REF_CODE_SHORT: {
-                    long distance = d.location - (r.location + 1);
+                    int distance = d.location - (r.location + 1);
                     
                     if (distance > 127 || distance < -128)
                         std::cerr << "REF_CODE_SHORT can't jump " << distance << " bytes!\n";
